tried to factor out tables in roles but the codebase was really busy so major refactoring is needed here
